readme intro-install-declarative
